Is an object created using tidyselect

choices and selected can be provided using tidyselect, (e.g. tidyselect::everything() tidyselect::where(), tidyselect::starts_with()). These functions can't be called independently but rather as an argument of function which consumes them. .is_tidyselect safely determines if x can be evaluated with tidyselect::eval_select()

Usage

.is_tidyselect(x)

Arguments

x

choices or selected

Value

logical(1)