Pick class constructor

Create a pick object

Usage

.pick(
  choices,
  selected,
  multiple = length(selected) > 1,
  ordered = FALSE,
  fixed = FALSE,
  ...
)

Arguments

choices

(tidyselect::language or character) Available values to choose.

selected

(tidyselect::language or character) Choices to be selected.

multiple

(logical(1)) if more than one selection is possible.

ordered

(logical(1)) if the selected should follow the selection order. If FALSE selected returned from srv_module_input() would be ordered according to order in choices.

fixed

(logical(1)) selection will be fixed and not possible to change interactively.

...

for picks(...): hierarchical structure that contains datasets() as first element and optionally variables() and values()

for variables(...) and values(...): additional arguments delivered to pickerInput, see shinyWidgets::pickerOptions() for available options as well as documentation for bootstrap-select v1.14.0-beta3 or higher for newer options (e.g., allow-clear that allows clearing the selection).

Value

pick generic object that is used by datasets(), variables() and values() to create objects of corresponding classes.